ai_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.8.2
python-dotenv==1.0.1
huggingface-hub==0.31.2
